training/pyproject.toml
[project]
name = "unitree-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.5.0 ; sys_platform == 'linux'",
    "jax[cpu] ; sys_platform == 'win32'",
    "jax[cuda] ; sys_platform == 'linux'",
    "mlflow>=3.5.1",
    "warp-lang>=1.10.0",
    "mujoco-mjx>=3.3.7",
    "tqdm>=4.67.1",
    "hydra-core>=1.3.2",
    "flatten-dict>=0.4.2",
    "nvidia-ml-py>=13.580.82",
    "torchrl>=0.10.1",
    "psutil>=7.1.3",
    "azureml-mlflow>=1.60.0",
    "azure-ai-ml>=1.30.0",
    "azure-identity>=1.25.1",
    "onnxscript>=0.5.6",
]

[tool.uv]
required-environments = [
    "sys_platform == 'win32'",
    "sys_platform == 'linux'"
]

[tool.uv.sources]
torch = { index = "pytorch-cuda-121" }

[[tool.uv.index]]
name = "pytorch-cuda-121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[dependency-groups]
dev = [
    "jupyterlab>=4.4.10",
    "mediapy>=1.2.4",
    "mujoco-python-viewer>=0.1.4",
    "onnx>=1.20.0",
]
